{"schema":"decodex/commit/2","change":"Realign close-on-exec test with ProcessPlatform ownership","authority":"manual","impact":"compatible"}
